pyproject.toml
[project]
name = "flask-noticed"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Webster Avosa", email = "websterb17@gmail.com" }
]
dependencies = [
    "flask==3.0.2",
    "sqlalchemy>=2.0.31",
    "alembic>=1.13.2",
    "celery>=5.4.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.0.7",
    "flask-socketio>=5.3.6",
    "requests>=2.32.3",
    "firebase-admin>=6.5.0",
    "twilio>=9.2.3",
    "vonage>=3.16.0",
    "pyjwt>=2.8.0",
    "cryptography>=43.0.0",
    "python-dotenv>=1.0.1",
    "redis>=5.0.7",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "pytest-flask>=1.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/flask_noticed"]
